Added the server code for project 5
Doesn't have any gui elements. I anticipate we will need at the very least a button that allows us to start the game once there are at least 4 players and 2 text fields. One that tells us number of clients in the server and the other one that tells us what the hint will be for the client based on the number they put in. This should be the bare minimum for the gui. We could have it look nicer though.